% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMouse-functions.R
\name{coord}
\alias{coord}
\title{Get Coordinates of Mouse Cursor}
\usage{
coord()
}
\value{
Returns a list with the coordinates.
}
\description{
Get the coordinates of the current location of the mouse cursor.
}
\examples{
\dontrun{
coord()
coord()$x
coord()$y
}
}
\seealso{
\code{\link{pos}}, \code{\link{record}}
}
\keyword{coord,}
\keyword{coordinates}
