% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMouse-functions.R
\name{delay}
\alias{delay}
\title{Time Delay}
\usage{
delay(ms = 500, failSafe = TRUE)
}
\arguments{
\item{ms}{(milliseconds). Default is \code{500}.}

\item{failSafe}{Logical flag to allow for emergency stops. Default is \code{TRUE}.}
}
\description{
Set the time delay in milliseconds manually.
}
\note{
If failSafe is \code{TRUE}, the function checks whether the mouse cursor is at the upper left position of the screen (pixel: x = 0, y = 0) . If so, the function fails and throws an error, allowing for emergency stops during code execution.
}
\examples{
\dontrun{
delay(2000)   # wait 2 seconds
move(0,0)
}
}
\seealso{
\code{\link{setAutoDelay}}
}
\keyword{delay}
