% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMouse-functions.R
\name{pos}
\alias{pos}
\title{Current position of mouse cursor}
\usage{
pos()
}
\value{
Prints "move(x,y)" to the console.
}
\description{
Print the position of the current location of the mouse cursor.
}
\note{
The console output can be pasted to a script, allowing to write mouse macros faster.
}
\examples{
\dontrun{
pos()
}
}
\seealso{
\code{\link{coord}}, \code{\link{record}}
}
\keyword{coordinates}
\keyword{position,}
