% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMouse-functions.R
\name{setAutoDelay}
\alias{setAutoDelay}
\title{Set automatic delay between commands}
\usage{
setAutoDelay(ms = 100)
}
\arguments{
\item{ms}{(milliseconds). Default is \code{100}.}
}
\description{
Set the time delay in milliseconds between commands.
}
\examples{
\dontrun{
setAutoDelay()
setAutoDelay(1000)   # 1 Second between commands
move(100, 100)
move(50, 50)
}
}
\seealso{
\code{\link{delay}}
}
\keyword{setAutoDelay}
